Add logging infrastructure
This class provides a relatively simple set of utility functions for
logging intermediate data, either to the screen or to MySQL. Currently
includes functions for intermediate data a la what delay-servo
produces and also the bit-rates that controlled-delay calculates.

This introduces a dependency for a file called mysql-credentials.hh
that should include the necessary MySQL information (all the undefined
constants used in the Logger() constructor). This file should not be
added to Git (this commit adds it to .gitignore). Some day, this
should really use ~/.my.cnf or something.